Add files via upload
Fixed error where event was getting 0 as event unit. This was because I need to keep the res = client.get_ts() before event to clear buffer. Otherwise extra strobe events will get added.
Fixed error test in which psth was clearing the unit_dict before decoding which caused all of the tilts to be classified as the same event.