Add files via upload
Changed how unit_dict_templates are created so that they are not just referencing the variable. 
Fixed a line in def psthtemplate() which was likely causing the templates to be much lower values than expected.
Currently working to fix a bug when number of units is change (lowered?) between recordings.